pyproject.toml
[tool.poetry]
name = "alphaservice"
version = "0.1.0"
description = ""
authors = ["GAU <gau@1jz.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
django-bootstrap5 = "^23.3"
pillow = "^10.1.0"
djangorestframework = "^3.14.0"
djangorestframework-xml = "^2.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
